third_party/blink/web_tests/external/wpt/css/cssom/idlharness-expected.txt

This is a testharness.js-based test.
Found 57 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSSStyleRule interface: existence and properties of interface object
  assert_equals: prototype of CSSStyleRule is not CSSGroupingRule expected function "function CSSGroupingRule() { [native code] }" but got function "function CSSRule() { [native code] }"
[FAIL] CSSStyleRule interface: existence and properties of interface prototype object
  assert_equals: prototype of CSSStyleRule.prototype is not CSSGroupingRule.prototype expected object "[object CSSGroupingRule]" but got object "[object CSSRule]"
[FAIL] CSSImportRule interface: sheet.cssRules[0] must inherit property "layerName" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] CSSImportRule interface: sheet.cssRules[0] must inherit property "supportsText" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] CSSPageDescriptors interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface object length
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface object name
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginTop
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginRight
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginBottom
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginLeft
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-top
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-right
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-bottom
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-left
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute size
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute pageOrientation
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute page-orientation
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marks
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute bleed
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors must be primary interface of sheet.cssRules[2].style
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] Stringification of sheet.cssRules[2].style
  assert_class_string: class string of sheet.cssRules[2].style expected "[object CSSPageDescriptors]" but got "[object CSSStyleDeclaration]"
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "margin" with the proper type
  assert_inherits: property "margin" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "marginTop" with the proper type
  assert_inherits: property "marginTop" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "marginRight" with the proper type
  assert_inherits: property "marginRight" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "marginBottom" with the proper type
  assert_inherits: property "marginBottom" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "marginLeft" with the proper type
  assert_inherits: property "marginLeft" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "margin-top" with the proper type
  assert_inherits: property "margin-top" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "margin-right" with the proper type
  assert_inherits: property "margin-right" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "margin-bottom" with the proper type
  assert_inherits: property "margin-bottom" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "margin-left" with the proper type
  assert_inherits: property "margin-left" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "size" with the proper type
  assert_inherits: property "size" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "pageOrientation" with the proper type
  assert_inherits: property "pageOrientation" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "page-orientation" with the proper type
  assert_inherits: property "page-orientation" found on object expected in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "marks" with the proper type
  assert_inherits: property "marks" not found in prototype chain
[FAIL] CSSPageDescriptors interface: sheet.cssRules[2].style must inherit property "bleed" with the proper type
  assert_inherits: property "bleed" not found in prototype chain
[FAIL] CSSStyleProperties interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface object length
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface object name
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: attribute cssFloat
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties must be primary interface of sheet.cssRules[4].style
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] Stringification of sheet.cssRules[4].style
  assert_class_string: class string of sheet.cssRules[4].style expected "[object CSSStyleProperties]" but got "[object CSSStyleDeclaration]"
[FAIL] CSSStyleProperties must be primary interface of sheet.cssRules[2].cssRules[0].style
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] Stringification of sheet.cssRules[2].cssRules[0].style
  assert_class_string: class string of sheet.cssRules[2].cssRules[0].style expected "[object CSSStyleProperties]" but got "[object CSSStyleDeclaration]"
[FAIL] CSSStyleProperties must be primary interface of style_element.style
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] Stringification of style_element.style
  assert_class_string: class string of style_element.style expected "[object CSSStyleProperties]" but got "[object CSSStyleDeclaration]"
[FAIL] CSSStyleProperties must be primary interface of svg_element.style
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] Stringification of svg_element.style
  assert_class_string: class string of svg_element.style expected "[object CSSStyleProperties]" but got "[object CSSStyleDeclaration]"
[FAIL] CSSStyleProperties must be primary interface of getComputedStyle(svg_element)
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] Stringification of getComputedStyle(svg_element)
  assert_class_string: class string of getComputedStyle(svg_element) expected "[object CSSStyleProperties]" but got "[object CSSStyleDeclaration]"
Harness: the test ran to completion.
